Proposed package refactor. Added *.iml to .gitignore.
