+++ src/main/java/co/alphavantage/input/timeseries/OutputSize.java
@@ -1,0 +1,27 @@
+package co.alphavantage.input.timeseries;
+
+import co.alphavantage.input.ApiParameter;
+
+/**
+ * Outputsize parameter, i.e outputsize=compact
+ */
+public enum OutputSize implements ApiParameter {
+  COMPACT("compact"),
+  FULL("full");
+
+  private final String urlParameter;
+
+  OutputSize(String urlParameter) {
+    this.urlParameter = urlParameter;
+  }
+
+  @Override
+  public String getKey() {
+    return "outputsize";
+  }
+
+  @Override
+  public String getValue() {
+    return urlParameter;
+  }
+}
