+++ src/main/java/co/alphavantage/output/AlphaVantageException.java
@@ -1,0 +1,11 @@
+package co.alphavantage.output;
+
+public class AlphaVantageException extends Exception {
+  public AlphaVantageException(String message, Exception e) {
+    super(message, e);
+  }
+
+  public AlphaVantageException(String message) {
+    super(message);
+  }
+}
