--- src/main/java/input/technical_indicator/Function.java
@@ -1,23 +1,0 @@
-package input.technical_indicator;
-
-import input.ApiParameter;
-
-public enum Function implements ApiParameter {
-  SMA("SMA");
-
-  private final String urlParameter;
-
-  Function(String urlParameter) {
-    this.urlParameter = urlParameter;
-  }
-
-  @Override
-  public String getKey() {
-    return "function";
-  }
-
-  @Override
-  public String getValue() {
-    return urlParameter;
-  }
-}
