--- src/main/java/input/technical_indicator/Interval.java
@@ -1,30 +1,0 @@
-package input.technical_indicator;
-
-import input.ApiParameter;
-
-public enum Interval implements ApiParameter {
-  ONE_MIN("1min"),
-  FIVE_MIN("5min"),
-  FIFTEEN_MIN("15min"),
-  THIRTY_MIN("30min"),
-  SIXTY_MIN("60min"),
-  DAILY("daily"),
-  WEEKLY("weekly"),
-  MONTHLY("monthly");
-
-  private final String urlParameter;
-
-  Interval(String urlParameter) {
-    this.urlParameter = urlParameter;
-  }
-
-  @Override
-  public String getKey() {
-    return null;
-  }
-
-  @Override
-  public String getValue() {
-    return null;
-  }
-}
