--- src/main/java/input/technical_indicator/SeriesType.java
@@ -1,26 +1,0 @@
-package input.technical_indicator;
-
-import input.ApiParameter;
-
-public enum SeriesType implements ApiParameter {
-  CLOSE("close"),
-  OPEN("open"),
-  HIGH("high"),
-  LOW("low");
-
-  private final String urlParameter;
-
-  SeriesType(String urlParameter) {
-    this.urlParameter = urlParameter;
-  }
-
-  @Override
-  public String getKey() {
-    return null;
-  }
-
-  @Override
-  public String getValue() {
-    return null;
-  }
-}
