--- src/main/java/input/time_series/Function.java
@@ -1,32 +1,0 @@
-package input.time_series;
-
-import input.ApiParameter;
-
-/**
- * Function parameter, i.e function=TIME_SERIES_INTRADAY
- */
-public enum Function implements ApiParameter {
-  INTRADAY("TIME_SERIES_INTRADAY"),
-  DAILY("TIME_SERIES_DAILY"),
-  DAILY_ADJUSTED("TIME_SERIES_DAILY_ADJUSTED"),
-  WEEKLY("TIME_SERIES_WEEKLY"),
-  WEEKLY_ADJUSTED("TIME_SERIES_WEEKLY_ADJUSTED"),
-  MONTHLY("TIME_SERIES_MONTHLY"),
-  MONTHLY_ADJUSTED("TIME_SERIES_MONTHLY_ADJUSTED");
-
-  private final String urlParameter;
-
-  Function(String urlParameter) {
-    this.urlParameter = urlParameter;
-  }
-
-  @Override
-  public String getKey() {
-    return "function";
-  }
-
-  @Override
-  public String getValue() {
-    return this.urlParameter;
-  }
-}
