--- src/main/java/input/time_series/OutputSize.java
@@ -1,27 +1,0 @@
-package input.time_series;
-
-import input.ApiParameter;
-
-/**
- * Outputsize parameter, i.e outputsize=compact
- */
-public enum OutputSize implements ApiParameter {
-  COMPACT("compact"),
-  FULL("full");
-
-  private final String urlParameter;
-
-  OutputSize(String urlParameter) {
-    this.urlParameter = urlParameter;
-  }
-
-  @Override
-  public String getKey() {
-    return "outputsize";
-  }
-
-  @Override
-  public String getValue() {
-    return urlParameter;
-  }
-}
