--- src/main/java/output/AlphaVantageException.java
@@ -1,11 +1,0 @@
-package output;
-
-public class AlphaVantageException extends Exception {
-  public AlphaVantageException(String message, Exception e) {
-    super(message, e);
-  }
-
-  public AlphaVantageException(String message) {
-    super(message);
-  }
-}
